Added a Guest service class

+++ Program.cs
@@ -4,6 +4,7 @@ using DB_HotelBooking1.Models;
 using System;
 using System.Linq;
 using DB_HotelBooking1.Services;
+
 namespace DB_HotelBooking1
 {
     internal class Program
@@ -11,10 +12,11 @@ namespace DB_HotelBooking1
         static void Main(string[] args)
         {
             var options = new DbContextOptionsBuilder<HotelContext>()
-       .UseSqlServer("Server=GABRIEL_U;Database=DB_HotelBooking1;Trusted_Connection=True;TrustServerCertificate=True;")
+                .UseSqlServer("Server=GABRIEL_U;Database=DB_HotelBooking1;Trusted_Connection=True;TrustServerCertificate=True;")
                 .Options;
 
-            using (var context = new HotelContext(options)) 
+            // Testing connection: list all rooms from the context
+            using (var context = new HotelContext(options))
             {
                 var rooms = context.Rooms.ToList();
                 foreach (var room in rooms)
@@ -24,67 +26,97 @@ namespace DB_HotelBooking1
             }
 
             Console.WriteLine("Testing");
-       
-        var hotelService = new HotelServices();
-        bool exit = false;
 
+           
+            var hotelService = new HotelServices();
+            bool exit = false;
 
-        while (!exit)
+            while (!exit)
             {
-                Console.WriteLine("\n --- Hotel Managment System ---");
+                Console.WriteLine("\n --- Hotel Management System ---");
                 Console.WriteLine("1. Add a room");
                 Console.WriteLine("2. List all rooms");
                 Console.WriteLine("3. Update a room");
                 Console.WriteLine("4. Delete a room");
                 Console.WriteLine("5. Exit");
-                Console.WriteLine("Choose an option: ");
-
+                Console.Write("Choose an option: ");
                 string? option = Console.ReadLine();
 
                 switch (option)
-
                 {
                     case "1":
-                        Console.WriteLine("Enter room type: ");
-                        string roomType = Console.ReadLine();
-                        Console.WriteLine("Enter number of extra beds: ");
-                        int extraBeds = int.Parse(Console.ReadLine());
+                        Console.Write("Enter room type: ");
+                        string? roomType = Console.ReadLine();
+                        if (string.IsNullOrWhiteSpace(roomType))
+                        {
+                            Console.WriteLine("Room type cannot be empty.");
+                            break;
+                        }
+
+                        Console.Write("Enter number of extra beds: ");
+                        string? extraBedsInput = Console.ReadLine();
+                        if (!int.TryParse(extraBedsInput, out int extraBeds))
+                        {
+                            Console.WriteLine("Invalid number for extra beds.");
+                            break;
+                        }
+
+                       
                         Room room = new Room(roomType, extraBeds);
                         hotelService.AddRoom(room);
                         break;
+
                     case "2":
                         hotelService.ListRooms();
                         break;
+
                     case "3":
-                        Console.WriteLine("Enter the room ID to update");
-                        int id = int.Parse(Console.ReadLine());
-                        Console.WriteLine("Enter new room type: ");
-                        string newRoomType = Console.ReadLine();
-                        Console.WriteLine("Enter new number of extra beds: ");
-                        int newExtraBeds = int.Parse(Console.ReadLine());
+                        Console.Write("Enter the room ID to update: ");
+                        string? idInput = Console.ReadLine();
+                        if (!int.TryParse(idInput, out int id))
+                        {
+                            Console.WriteLine("Invalid room ID.");
+                            break;
+                        }
+                        Console.Write("Enter new room type: ");
+                        string? newRoomType = Console.ReadLine();
+                        if (string.IsNullOrWhiteSpace(newRoomType))
+                        {
+                            Console.WriteLine("New room type cannot be empty.");
+                            break;
+                        }
+                        Console.Write("Enter new number of extra beds: ");
+                        string? newExtraBedsInput = Console.ReadLine();
+                        if (!int.TryParse(newExtraBedsInput, out int newExtraBeds))
+                        {
+                            Console.WriteLine("Invalid number for extra beds.");
+                            break;
+                        }
                         hotelService.UpdateRoom(id, newRoomType, newExtraBeds);
                         break;
+
                     case "4":
-                        Console.WriteLine("Enter the room ID to delete");
-                        int deleteId = int.Parse(Console.ReadLine());
+                        Console.Write("Enter the room ID to delete: ");
+                        string? deleteIdInput = Console.ReadLine();
+                        if (!int.TryParse(deleteIdInput, out int deleteId))
+                        {
+                            Console.WriteLine("Invalid room ID.");
+                            break;
+                        }
                         hotelService.DeleteRoom(deleteId);
                         break;
+
                     case "5":
                         exit = true;
                         break;
+
                     default:
-                        Console.WriteLine("Invalid option. ");
+                        Console.WriteLine("Invalid option. Please try again.");
                         break;
-
-
                 }
-                
+            }
 
-             }
-
-            Console.WriteLine("Exiting the Application. Goodye! ");
-
+            Console.WriteLine("Exiting the Application. Goodbye!");
         }
     }
 }
-
